% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COD-ANCOVA.R
\name{ANCOVA}
\alias{ANCOVA}
\title{Información del paquete ANCOVA}
\usage{
ANCOVA(y, tratamiento, bloque, covariable, significancia, data)
}
\arguments{
\item{y}{(vector) nombre de la variable respuesta.}

\item{tratamiento}{(string) nombre de la variable que representa a los tratamientos.}

\item{bloque}{(string) nombre de la variable que representa los bloques.}

\item{covariable}{(vector) nombre de la covariable.}

\item{significancia}{(numeric) El valor de alfa o el valor de cuanto error se permite para el experimento}

\item{data}{Nombre de su archivo donde alberga la información la cual tiene que estar perfectamente acomodada en una tabla de datos formato largo con los datos
de los tratamientos, bloques, covariable y de la variable respuesta.}
}
\value{
Devuelve una tabla con los datos de la suma de cuadrados y productos cruzados de x,y,xy para los tratamientos, bloques, error, totales y error ajustado, tambien devuelve la F calculada de la covariable y de los tratamientos, la F tabulada de la covariable y de los tratamientos, ademas le proporciona el dato de coeficiente de covarianza para que en el caso de que su covariable sea significativa pueda hacer el ajuste de medias correspondiente.
}
\description{
El paquete ANCOVA tiene la función ANCOVA: Esta función le ayudará a obtener una tabla resumen del Análisis de covarianza (ANCOVA)
para un Diseño en Bloques Completamente al Azar.
Esto le permitirá conocer que tanto ha influido una covariable en la variable respuesta de
su diseño experimental.
}
\details{
Para comenzar usted necesita tener sus datos acomodados en formato largo ya sea en Excel, csv, data.frame, u otra aplicación
que Rstudio le permita cargar, asegurese de que sus tratamientos o bloques sean nombrados con datos numericos; una vez cubierta esa parte, tendrá que cargar la función en su ordenador junto con el paquete Knitr
(este ultimo ayudará a la función a ordenar sus resultados en una tabla) la función ya cargará el knitr por usted: "install.packages("knitr")"
}
\examples{

# A continuación se muestra un ejemplo para comprender el uso de la función que alberga el paquete ANCOVA:

En un cultivo de sorgo forrajero se evaluó el rendimiento de grano y
la producción de forraje. Para ello, se compararon 3 tratamientos: 30, 40, y 50 días
después de ocurrida la floración. El número de plantas por parcela útil fue de 52.
A la cosecha el número de plantas por tratamiento y repetición fue diferente.
Los datos obtenidos al finalizar el experimento se presentan en el siguiente cuadro,
en el cual X representa el número de plantas por parcela y Y la producción
en grano en kilogramos por parcela. El diseño experimental.

DATOS:
 df <- data.frame(bloque = c(1,1,1,1,1,1,,2,2,2,2,2,2,3,3,3,3,3,3),
                  tratamiento = c(1,1,1,1,1,1,2,2,2,2,2,2,3,3,3,3,3,3),
                  Nplantas = c(41,24,31,46,32,38,37,32,38,41,42,40,37,34,47,41,50,39),
                  producción = c(4.08,2.78, 2.79, 4.24, 4.17, 2.62, 4.72, 4.92, 4.5,
                                 5.62, 5.15, 4.32, 4, 5.05, 5.54, 6.46, 6.65, 5.7))

Significancia de 0.05

# Paso 1: Limpiar la memoria de R con el comando

rm(list=ls())

# Paso 2: Cargamos el paquete ANCOVA:

library(ANCOVA)


# la función pide los argumentos siguientes

ANCOVA(y, tratamiento, bloque, covariable, significancia, data)

# Para este caso tendriamos que argumentar a la función así:

 ANCOVA(y="producción", tratamiento = "tratamiento", bloque = "bloque",
        covariable = "Nplantas", significancia = 0.05, data = df)

# O simplemente puede:

 ANCOVA("producción", "tratamiento","bloque", "Nplantas", 0.05, df)


# Paso 3: Usted tendrá consigo dos tablas del análisis de covarianza de sus datos:
Una tabla será de la suma de cuadrados y la otra de las F
calculadas y tabuladas tanto para la covarianza como  para el tratamiento.

CON DICHAS TABLAS PODRÁ HACER DEDUCCIONES DE QUE TAN INFLUYENTE RESULTÓ SU COVARIABLE EN
LOS RESULTADOS EXPERIMENTALES.

1. SI F_cal_trat > F_tab_trat, se rechaza Ho por lo que se deduce que la covariable
es significativa al menos en alguno de los bloques y por lo tanto se tendría que hacer un ajuste de medias.

2. SI F_cal_cov > F_tab_cov, NO se rechaza Ho por lo que la covariable en
su experimento no es significativa.


Las tablas de las F calculadas y tabuladas le anexará el coeficiente de covarianza (COEF_COVARIANZA) el cual servirá para hacer el ajuste de medias correspondiente.

Para hacer este ajuste de medias usted puede usar la función "ANCOVA_AJUST" que al alberga este paquete, para más información puede consultar el manual.
}
